fix update menu

+++ update_menu.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 set +x
 
 # Update menu
@@ -9,9 +9,6 @@ set +x
 
 # use wget
 # bash <(wget -qO- https://raw.githubusercontent.com/YogSottot/DebianLikeBitrixVM/feature/php-fpm/update_menu.sh)
-
-cat > /root/temp_update_menu.sh <<\END
-#!/usr/bin/env bash
 
 BRANCH="feature/php-fpm"
 REPO_URL="https://github.com/YogSottot/DebianLikeBitrixVM.git"
@@ -37,21 +34,24 @@ cd "$DEST_DIR_MENU/DebianLikeBitrixVM"
 git sparse-checkout set $DIR_NAME_MENU
 
 # Move vm_menu in /root and clean
-rm -rf $DEST_DIR_MENU/$DIR_NAME_MENU
+rm -rf "${DEST_DIR_MENU:?}/${DIR_NAME_MENU:?}"
 mv -f $DIR_NAME_MENU $DEST_DIR_MENU
-rm -rf "$DEST_DIR_MENU/DebianLikeBitrixVM"
+rm -rf "${DEST_DIR_MENU:?}/DebianLikeBitrixVM"
 
 chmod -R +x $DEST_DIR_MENU/$DIR_NAME_MENU
 
 # Check script in .profile and add to .profile if not exist
-if ! grep -qF "$FULL_PATH_MENU_FILE" /root/.profile; then
-  cat << INSTALL_MENU >> /root/.profile
+if [ "${BS_ADD_MENU_IN_BASH_PROFILE}" == 'Y' ]; then
+    # Check script in .profile and add to .profile if not exist
+    if ! grep -qF "$FULL_PATH_MENU_FILE" /root/.profile; then
+      cat << INSTALL_MENU >> /root/.profile
 
-#if [ -n "\$SSH_CONNECTION" ]; then
-#  $FULL_PATH_MENU_FILE
-#fi
+if [ -n "\$SSH_CONNECTION" ]; then
+    $FULL_PATH_MENU_FILE
+fi
 
 INSTALL_MENU
+    fi
 fi
 
 rm -f "/tmp/configs.tmp"
@@ -62,10 +62,3 @@ ln -s $FULL_PATH_MENU_FILE "$DEST_DIR_MENU/menu.sh"
 echo -e "\n\n";
 echo "Menu updated! Backup directory old menu: $full_path_backup_menu";
 echo -e "\n";
-
-END
-
-bash /root/temp_update_menu.sh
-
-rm /root/temp_update_menu.sh
-
